pyproject.toml
[project]
name = "rpg-mt"
version = "0.1.0"
description = "rpg-mt"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.13",
    "httpx>=0.28.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-huggingface>=0.3.1",
    "langchain-openai>=0.3.30",
    "sentence-transformers>=5.1.0",
    "sse-starlette>=3.0.2",
    "weaviate-client>=4.16.7",
]

[[tool.uv.index]]
name = "tuna"
url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple/"
default = true

[dependency-groups]
dev = []
